HW6/requirements.txt
pytorch_lightning==1.3.6
torchmetrics==0.3.2
mlflow==1.18.0
hydra_core==1.1.0
omegaconf==2.1.0
torch==1.9.0
numpy==1.20.3
